improvement(users): Separates user creating logic
