+++ packages/listener-solidarity/src/graphql/mutations/updateFormEntries.ts
@@ -12,7 +12,6 @@ const FORM_ENTRIES_MUTATION = gql`
     ) {
       returning {
         id
-        updated_at
       }
     }
   }
