+++ packages/listener-solidarity/src/utils/index.ts
@@ -1,3 +1,11 @@
+import { User } from "../types";
+
+export const organizationsIds = {
+  MSR: 360273031591,
+  THERAPIST: 360282119532,
+  LAWYER: 360269610652,
+};
+
 export const setType = (type: string | null) => {
   switch (type) {
     case "Acolhimento Jurídico":
@@ -21,6 +29,15 @@ export const getOrganizationType = (id: number): string => {
   if (therapist_widgets.includes(id)) return "THERAPIST";
   if (lawyer_widgets.includes(id)) return "LAWYER";
   return "MSR";
+};
+
+export const handleError = (entries: User[]) => {
+  console.log(
+    `Integration failed in these form entries ${entries.map(
+      (e) => e.external_id
+    )}`
+  );
+  return undefined;
 };
 
 export { default as getGeocoding } from "./geocoding";
